Separate tests that rely on SecurityManager (Jsr292)
SecurityManager will be removed in a future jdk. Separate tests that
rely on it so that they can be easily disabled in the future.

Issue: #14412
