+++ test/functional/Jsr292/testng_8.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <!--
-  Copyright (c) 2016, 2018 IBM Corp. and others
+  Copyright (c) 2016, 2022 IBM Corp. and others
 
   This program and the accompanying materials are made available under
   the terms of the Eclipse Public License 2.0 which accompanies this
@@ -51,6 +51,12 @@
 			<class name="com.ibm.j9.jsr292.GuardTest"/>
 			<class name="com.ibm.j9.jsr292.FilterArgumentsTest"/>
 			<class name="com.ibm.j9.jsr292.RestrictReceiverTest"/>
+		</classes>
+	</test>
+
+	<test name="jsr292Test_SM">
+		<classes>
+			<class name="com.ibm.j9.jsr292.MethodTypeTests_SM"/>
 		</classes>
 	</test>
